Move visibility logic to its own file.
